Fixed attribute change position processing
Previously, the change position byte was not processed correctly unless a value of 1 or more was set.
It also was not processed correctly if the values ​​were in reverse order.

Solution:
Read all change position bytes in advance and convert them to 80-bit flags.
The attribute is changed if a flag is set for the next column position to be processed.