Add mouse support
Supports PC-8872 compatible mouse. (aka "bus mouse")
The mouse communicates using a protocol compatible with MSX mouse.
This implementation is based on the ps2mouse module of MiSTer MSX core.

To enable the mouse, select it from the "Input" in the F12 menu.
(Default: Joypad)
The mouse cannot be used at the same time as the joypad.

NOTE:
The joystick mode on mouse is not support.