add RD_n to JT08
Change the trigger for starting a RAM WRITE/READ sequence from CS_n to WR_n/RD_n in jt08_adpcm_drvB.
Add the external input RD_n, which was omitted because it was not used until now.

RD_n is only used by JT08, so I won't add it to other wrappers.