Improved A1 input match check when writing to a register (JT12)
A1 is saved when selecting a register (A0=0;WRITE).
When register writing (A0=1;WRITE), it is used to determine the destination register.

If A1=0 when selecting a register and A1=1 when writing, the data was written to the same register with A1=0 as when selected.
This case is generally a software bug.
But from a hardware point of view, it is correct not to write to a register unless it is the same A1 that was selected.

[Solution]
When writing to a register, write data only if A1 is the same as when selected.