Modify the behavior of multiple interrupts (INTC)
In the uPD8214, "interrupt disable flip-flop" is set at the same time as INT is asserted, and the next INT is not asserted until it is reset.
INTC has been modified to comply with this specification.

This fixes an issue where the CPU was interrupted even though the interrupt handler implementation did not expect multiple interrupts.