Fix the digital palette behavior
Palette mode has been added to the conditions for determining whether palette colors are valid.

Adding a DMA stop with digital palette and B&W graphics mode settings means that the bits on the graphics plane are no longer colored with the palette colors.
This behavior differs from that of analog palette mode.